+++ subprojects/library/src/commonMain/kotlin/org/kotools/types/number/Integer.kt
@@ -1,0 +1,11 @@
+package org.kotools.types.number
+
+import org.kotools.types.ExperimentalKotoolsTypesApi
+
+/**
+ * Represents a mathematical integer, with exact operations that never overflow.
+ *
+ * @since 5.2.0
+ */
+@ExperimentalKotoolsTypesApi
+public class Integer private constructor()